Switched helper files
* helper.php -> legacy\helper.php
* helper2.php -> helper.php

Seems like the cleanest solution to have it like this. This way we can
hopefully slowly migrate more and more of the files_sharing app to our
new fancy frameworks with unit tests
