+++ apps/files_sharing/ajax/external.php
@@ -30,7 +30,7 @@ OCP\JSON::checkAppEnabled('files_sharing');
 $l = \OC::$server->getL10N('files_sharing');
 
 // check if server admin allows to mount public links from other servers
-if (OCA\Files_Sharing\Helper::isIncomingServer2serverShareEnabled() === false) {
+if (OCA\Files_Sharing\Legacy\Helper::isIncomingServer2serverShareEnabled() === false) {
 	\OCP\JSON::error(array('data' => array('message' => $l->t('Server to server sharing is not enabled on this server'))));
 	exit();
 }
