+++ apps/files_sharing/settings-admin.php
@@ -25,7 +25,7 @@
 \OCP\Util::addScript('files_sharing', 'settings-admin');
 
 $tmpl = new OCP\Template('files_sharing', 'settings-admin');
-$tmpl->assign('outgoingServer2serverShareEnabled', OCA\Files_Sharing\Helper::isOutgoingServer2serverShareEnabled());
-$tmpl->assign('incomingServer2serverShareEnabled', OCA\Files_Sharing\Helper::isIncomingServer2serverShareEnabled());
+$tmpl->assign('outgoingServer2serverShareEnabled', OCA\Files_Sharing\Legacy\Helper::isOutgoingServer2serverShareEnabled());
+$tmpl->assign('incomingServer2serverShareEnabled', OCA\Files_Sharing\Legacy\Helper::isIncomingServer2serverShareEnabled());
 
 return $tmpl->fetchPage();
